Refactor packages of model
