+++ src/main/java/com/jalasoft/jfc/model/image/ImageParam.java
@@ -5,9 +5,10 @@
  *  ("Confidential Information"). You shall not disclose such Confidential
  *  Information and shall use it only in accordance with the terms of the
  *  license agreement you entered into with Jalasoft.
+ *
  */
 
-package com.jalasoft.jfc.model;
+package com.jalasoft.jfc.model.image;
 
 /**
  * Manages params of an image.
